Move classes to appropriate packages
