+++ radixdlt/src/integration/java/com/radixdlt/consensus/deterministic/ControlledNetwork.java
@@ -21,8 +21,8 @@ import com.google.common.collect.ImmutableList;
 import com.radixdlt.EpochChangeSender;
 import com.radixdlt.consensus.bft.BFTEventReducer.BFTEventSender;
 import com.radixdlt.consensus.CommittedStateSync;
-import com.radixdlt.consensus.EpochChange;
-import com.radixdlt.consensus.LocalTimeout;
+import com.radixdlt.consensus.epoch.EpochChange;
+import com.radixdlt.consensus.epoch.LocalTimeout;
 import com.radixdlt.consensus.bft.BFTNode;
 import com.radixdlt.consensus.bft.GetVerticesErrorResponse;
 import com.radixdlt.consensus.bft.GetVerticesResponse;
