+++ radixdlt/src/integration/java/com/radixdlt/consensus/deterministic/ControlledNode.java
@@ -23,13 +23,13 @@ import com.radixdlt.consensus.bft.BFTNode;
 import com.radixdlt.consensus.CommittedStateSync;
 import com.radixdlt.consensus.ConsensusEvent;
 import com.radixdlt.consensus.DefaultHasher;
-import com.radixdlt.consensus.EmptySyncEpochsRPCSender;
-import com.radixdlt.consensus.EpochChange;
-import com.radixdlt.consensus.EpochManager;
+import com.radixdlt.consensus.epoch.EmptySyncEpochsRPCSender;
+import com.radixdlt.consensus.epoch.EpochChange;
+import com.radixdlt.consensus.epoch.EpochManager;
 import com.radixdlt.consensus.PendingVotes;
 import com.radixdlt.consensus.bft.GetVerticesErrorResponse;
 import com.radixdlt.consensus.bft.GetVerticesResponse;
-import com.radixdlt.consensus.LocalTimeout;
+import com.radixdlt.consensus.epoch.LocalTimeout;
 import com.radixdlt.consensus.ProposerElectionFactory;
 import com.radixdlt.consensus.VertexMetadata;
 import com.radixdlt.consensus.bft.VertexStore.GetVerticesRequest;
