+++ radixdlt/src/integration/java/com/radixdlt/consensus/simulation/configuration/SingleEpochAlwaysSyncedStateComputer.java
@@ -17,7 +17,7 @@
 
 package com.radixdlt.consensus.simulation.configuration;
 
-import com.radixdlt.consensus.EpochChange;
+import com.radixdlt.consensus.epoch.EpochChange;
 import com.radixdlt.consensus.Vertex;
 import com.radixdlt.consensus.VertexMetadata;
 import com.radixdlt.consensus.bft.BFTNode;
