+++ radixdlt/src/integration/java/com/radixdlt/consensus/simulation/network/SimulationNodes.java
@@ -25,8 +25,8 @@ import com.radixdlt.consensus.ConsensusRunner;
 import com.radixdlt.consensus.ConsensusRunner.Event;
 import com.radixdlt.consensus.ConsensusRunner.EventType;
 import com.radixdlt.consensus.DefaultHasher;
-import com.radixdlt.consensus.EmptySyncVerticesRPCSender;
-import com.radixdlt.consensus.EpochManager;
+import com.radixdlt.consensus.epoch.EmptySyncVerticesRPCSender;
+import com.radixdlt.consensus.epoch.EpochManager;
 import com.radixdlt.consensus.EpochChangeRx;
 import com.radixdlt.consensus.Hasher;
 import com.radixdlt.middleware2.InternalMessagePasser;
