+++ radixdlt/src/main/java/com/radixdlt/CerberusModule.java
@@ -31,11 +31,11 @@ import com.radixdlt.consensus.CommittedStateSyncRx;
 import com.radixdlt.consensus.ConsensusRunner;
 import com.radixdlt.consensus.DefaultHasher;
 import com.radixdlt.consensus.EpochChangeRx;
-import com.radixdlt.consensus.EpochManager;
+import com.radixdlt.consensus.epoch.EpochManager;
 import com.radixdlt.consensus.ConsensusEventsRx;
 import com.radixdlt.consensus.PendingVotes;
 import com.radixdlt.consensus.SyncEpochsRPCRx;
-import com.radixdlt.consensus.EpochManager.SyncEpochsRPCSender;
+import com.radixdlt.consensus.epoch.EpochManager.SyncEpochsRPCSender;
 import com.radixdlt.consensus.SyncedStateComputer;
 import com.radixdlt.consensus.VertexStoreEventsRx;
 import com.radixdlt.middleware2.InternalMessagePasser;
