+++ radixdlt/src/main/java/com/radixdlt/consensus/ConsensusRunner.java
@@ -17,6 +17,7 @@
 
 package com.radixdlt.consensus;
 
+import com.radixdlt.consensus.epoch.EpochManager;
 import com.radixdlt.consensus.liveness.PacemakerRx;
 
 import com.radixdlt.utils.ThreadFactories;
