+++ radixdlt/src/main/java/com/radixdlt/consensus/epoch/EmptySyncEpochsRPCSender.java
@@ -15,18 +15,24 @@
  * language governing permissions and limitations under the License.
  */
 
-package com.radixdlt.consensus;
+package com.radixdlt.consensus.epoch;
 
-import static org.mockito.Mockito.mock;
-
+import com.radixdlt.consensus.VertexMetadata;
 import com.radixdlt.consensus.bft.BFTNode;
-import org.junit.Test;
 
-public class EmptySyncEpochsRPCSenderTest {
-	@Test
-	public void when_send_request_and_response__then_no_exception_occurs() {
-		EmptySyncEpochsRPCSender.INSTANCE.sendGetEpochRequest(mock(BFTNode.class), 12345L);
-		EmptySyncEpochsRPCSender.INSTANCE.sendGetEpochResponse(mock(BFTNode.class), mock(VertexMetadata.class));
+/**
+ * A mocked sync epochs rpc sender
+ */
+public enum EmptySyncEpochsRPCSender implements EpochManager.SyncEpochsRPCSender {
+	INSTANCE;
+
+	@Override
+	public void sendGetEpochRequest(BFTNode node, long epoch) {
+		// No-op
 	}
 
+	@Override
+	public void sendGetEpochResponse(BFTNode node, VertexMetadata ancestor) {
+		// No-op
+	}
 }
