+++ radixdlt/src/main/java/com/radixdlt/consensus/epoch/EpochChange.java
@@ -15,8 +15,9 @@
  * language governing permissions and limitations under the License.
  */
 
-package com.radixdlt.consensus;
+package com.radixdlt.consensus.epoch;
 
+import com.radixdlt.consensus.VertexMetadata;
 import com.radixdlt.consensus.bft.BFTValidatorSet;
 import java.util.Objects;
