+++ radixdlt/src/main/java/com/radixdlt/consensus/liveness/PacemakerRx.java
@@ -17,7 +17,7 @@
 
 package com.radixdlt.consensus.liveness;
 
-import com.radixdlt.consensus.LocalTimeout;
+import com.radixdlt.consensus.epoch.LocalTimeout;
 import io.reactivex.rxjava3.core.Observable;
 
 /**
